test: hardware listing
Closes #981
