+++ backend/kernelCI_app/unitTests/utils/fields/hardware.py
@@ -1,0 +1,10 @@
+hardware_listing_fields = [
+    "hardware_name",
+    "platform",
+    "test_status_summary",
+    "boot_status_summary",
+    "build_status_summary",
+]
+
+test_status_summary_fields = ["FAIL", "PASS", "SKIP", "ERROR", "MISS", "NULL", "DONE"]
+build_status_summary_fields = ["valid", "invalid", "null"]
